load SPLADE models in fp16

+++ langchain_websearch.py
@@ -5,6 +5,7 @@ import logging
 
 import aiohttp
 import requests
+import torch
 from bs4 import BeautifulSoup
 from langchain.retrievers.document_compressors import DocumentCompressorPipeline
 from langchain.retrievers.ensemble import EnsembleRetriever
@@ -18,7 +19,6 @@ from langchain_community.document_transformers import EmbeddingsRedundantFilter
 from langchain_community.retrievers import BM25Retriever
 from transformers import AutoTokenizer, AutoModelForMaskedLM
 import optimum.bettertransformer.transformation
-
 try:
     from qdrant_client import QdrantClient, models
 except ImportError:
@@ -47,11 +47,11 @@ class LangchainCompressor:
             self.splade_doc_tokenizer = AutoTokenizer.from_pretrained("naver/efficient-splade-VI-BT-large-doc",
                                                                       cache_dir=model_cache_dir)
             self.splade_doc_model = AutoModelForMaskedLM.from_pretrained("naver/efficient-splade-VI-BT-large-doc",
-                                                                         cache_dir=model_cache_dir).to(self.device)
+                                                                         cache_dir=model_cache_dir, torch_dtype=torch.float16).to(self.device)
             self.splade_query_tokenizer = AutoTokenizer.from_pretrained("naver/efficient-splade-VI-BT-large-query",
                                                                         cache_dir=model_cache_dir)
             self.splade_query_model = AutoModelForMaskedLM.from_pretrained("naver/efficient-splade-VI-BT-large-query",
-                                                                           cache_dir=model_cache_dir).to(self.device)
+                                                                           cache_dir=model_cache_dir, torch_dtype=torch.float16).to(self.device)
             optimum_logger = optimum.bettertransformer.transformation.logger
             original_log_level = optimum_logger.level
             # Set the level to 'ERROR' to ignore "The BetterTransformer padding during training warning"
@@ -153,7 +153,6 @@ class LangchainCompressor:
             weights=[self.ensemble_weighting, 1 - self.ensemble_weighting]
         )
         compressed_docs = ensemble_retriever.invoke(query)
-
         # Ensemble may return more than "num_results" results, so cut off excess ones
         return compressed_docs[:self.num_results]
